Determine folder exists

+++ ci/scripts/prepare_source_code_cicd.sh
@@ -9,6 +9,8 @@ VELOX_BRANCH_NAME=`git branch -r --contains ${VELOX_COMMIT_ID} | cut -d' ' -f3`
 echo "branch name: ${VELOX_BRANCH_NAME}"
 popd
 PRESTO_LOCAL_BRANCH_NAME=`echo ${VELOX_BRANCH_NAME} | cut -d '/' -f2`
-rm -rf $PRESTO_LOCAL_PATH
+if [ -d $PRESTO_LOCAL_PATH ]; then
+    rm -rf $PRESTO_LOCAL_PATH
+fi
 git clone -b $PRESTO_LOCAL_BRANCH_NAME https://github.com/intel-innersource/frameworks.ai.modular-sql.presto.git $PRESTO_LOCAL_PATH
 mkdir $PRESTO_LOCAL_PATH/presto-native-execution/BDTK
